math function 1 complete
got my first math function done. To be honest I ran into so many syntax
errors putting that small function together I’m not even 100% sure why
it worked in that incarnation. I’m going to look over it a few more
times and read back in the class material to figure out why it works
now and not on my earlier tries.